Split interfaces.ts into separate files
