+++ src/app/core/auth.service.ts
@@ -3,13 +3,13 @@ import { AngularFireAuth } from '@angular/fire/auth';
 import { AngularFirestore, AngularFirestoreDocument } from '@angular/fire/firestore';
 import { Router } from '@angular/router';
 import { take } from 'rxjs/operators';
-import { BetaUsers, VirtrolioUser } from '../shared/interfaces';
 import { Location } from '@angular/common';
 import { DeviceDetectorService } from 'ngx-device-detector';
 import firebase from 'firebase/app';
 import 'firebase/auth';
 import User = firebase.User;
 import auth = firebase.auth;
+import { BetaUsers, VirtrolioUser } from '../shared/interfaces/users';
 
 @Injectable({
   providedIn: 'root'
