+++ src/app/core/msg-io.service.ts
@@ -1,6 +1,5 @@
 import { Injectable, SecurityContext } from '@angular/core';
 import { AngularFirestore, AngularFirestoreCollection, AngularFirestoreDocument } from '@angular/fire/firestore';
-import { VirtrolioDocument, VirtrolioMessage, VirtrolioMessageTemplate } from '../shared/interfaces';
 
 import { map, take } from 'rxjs/operators';
 import { Observable } from 'rxjs';
@@ -12,6 +11,7 @@ import firebase from 'firebase/app';
 import 'firebase/firestore';
 import firestore = firebase.firestore;
 import Timestamp = firestore.Timestamp;
+import { VirtrolioDocument, VirtrolioMessage, VirtrolioMessageTemplate } from '../shared/interfaces/messages';
 
 @Injectable({
   providedIn: 'root'
@@ -78,7 +78,7 @@ export class MsgIoService {
 
   /**
    * Collects ALL messages from the Firestore messages database that were sent to a particular user.
-   * Pay careful attention to the fields that are returned in a VirtrolioMessage by reading interfaces.ts.
+   * Pay careful attention to the fields that are returned in a VirtrolioMessage by reading shared/interfaces/messages.ts.
    * A VirtrolioMessage is NOT identical to a VirtrolioMessageTemplate.
    * @returns An Observable that will contain an array of all messages sent to uid, including the message IDs.
    * @throws Error - If the argument is blank, null or undefined.
@@ -98,7 +98,7 @@ export class MsgIoService {
 
   /**
    * Collects a SINGLE message from the Firestore messages database based on a message ID.
-   * Pay careful attention to the fields that are returned in a VirtrolioMessage by reading interfaces.ts.
+   * Pay careful attention to the fields that are returned in a VirtrolioMessage by reading shared/interfaces/messages.ts.
    * A VirtrolioMessage is NOT identical to a VirtrolioMessageTemplate.
    * @param msgID - The ID of the message to be received.
    */
