+++ src/app/pages/viewing/responses-list/responses-list.component.ts
@@ -2,8 +2,8 @@ import { Component, Input, OnInit } from '@angular/core';
 import { ViewingService } from '../../../core/viewing.service';
 import { ViewportScroller } from '@angular/common';
 import { AuthService } from '../../../core/auth.service';
-import { VirtrolioMessage } from '../../../shared/interfaces';
 import { CookieService } from 'ngx-cookie-service';
+import { VirtrolioMessage } from '../../../shared/interfaces/messages';
 
 @Component({
   selector: 'app-responses-list',
