+++ src/app/shared/interfaces/faq.ts
@@ -1,0 +1,9 @@
+export interface FAQQuestion {
+    question: string;
+    answer: string;
+}
+
+export interface FAQSection {
+    sectionTitle: string;
+    questions: FAQQuestion[];
+}
