+++ src/app/shared/interfaces/users.ts
@@ -1,0 +1,9 @@
+export interface VirtrolioUser {
+    displayName: string;
+    key: string;
+    profilePic: string;
+}
+
+export interface BetaUsers {
+    users: string[];
+}
